requirements.txt
streamlit
firebase-admin
google-genai
openai
streamlit-cookies-controller